scripts/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
metaphor-python==0.1.11
multidict==6.0.4
openai==0.27.8
openai-function-call==0.2.2
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
regex==2023.8.8
requests==2.31.0
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
